services/login/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "login"
version = "0.1.0"
description = ""
authors = ["Ido Saban <sabanido.business@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.72.0"
uvicorn = { version = "^0.17.0", extras = ["standard"] }
pyhumps = "^3.5.0"
PyJWT = "^2.3.0"
motor = "^2.5.1"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
isort = "^5.10.1"
flake8 = "^4.0.1"
mypy = "^0.931"
pytest = "^6.2.5"
coverage = { version = "^6.2", extras = ["toml"] }
requests = "^2.27.0"
pytest-env = "^0.6.2"

[tool.black]
line-length = 88
target_version = ["py310"]

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.10"
plugins = ["pydantic.mypy"]

[tool.pydantic-mypy]
warn_untyped_fields = true

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = ["tests"]
# env = []

[tool.coverage.run]
source = ["flights"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover", # Re-enable the standard pragma
    "def __repr__", # Don't complain about missing debug-only code
    "raise NotImplementedError"
]
